Use Sidekiq::Process
This makes the code easier to read and relies on official Sidekiq APIs,
which are less likely to break. It also only looks at the current
process to see, whether it is still processing jobs, not any quieted
process.

`identity` is inherited from `Sidekiq::Util`.

However, it relies on APIs added in Sidekiq 5